fix(ops): chunked_sgmv_expand kunlunxin long-index fix (T47 e4)
Failure intel via the authenticated submissions API (raw_result carries
failed_cases the CLI status view strips): every kunlunxin failure across
E2/E3 was 'index_copy_(): Expected a long tensor for index, but got
Int' - the platform's permutation arrives as int32 and kunlunxin torch
requires long indices (NVIDIA accepts int32, so the proxy variants
matrix stayed green). One-line fix: rows.long() before the route ops.

+++ src/flaggems_sglang/runtime/backend/_kunlunxin/ops/chunked_sgmv_expand.py
@@ -149,7 +149,10 @@ def chunked_sgmv_expand(
         if lora_ranks[w_idx] == 0:
             continue
         scaling = float(scalings[w_idx])
-        rows = permutation[start:end]
+        # the platform hands permutation as int32; kunlunxin torch
+        # requires long indices for index_select/index_copy_ (NVIDIA
+        # accepts int32, which is why the proxy stayed green)
+        rows = permutation[start:end].long()
         x_seg = x.index_select(0, rows).float()
         seg_out = output.index_select(0, rows).float()
         for i in range(n_slices):
